// src/app/models/shipment.ts
export interface Shipment {
    shipment_id: string
    booking_number: string
    contractual_customer_id: string
    transport_start_date: string
    create_date: string
    shipment_status: string
    brand: string
    is_priced_by_afls: boolean
    shipment_state_event_comment_text: any
    state_event_define: any
    state_event_detail_define: any
    receipt_mode_type: string
    delivery_mode_type: string
    route_code: string
    trade_code: string
    trade_lane_code: string
    trade_lane_name: string
    trade_lane_is_maersk_us_exports: boolean
    origin_location_id: string
    origin_location_name: string
    destination_location_id: string
    destination_location_name: string
    origin_departure_time_utc_expected: string
    origin_departure_date: string
    origin_departure_delay: number
    n_operational_route_points: number
    operational_route_transport_sequence: string
    destination_arrival_time_utc_expected: string
    destination_arrival_date: string
    destination_arrival_delay: number
    delay_bucket: string
    lopfi_location_id: string
    lopfi_location_name: string
    lopfi_voyage: string
    lopfi_service: string
    lopfi_vessel_code: string
    lopfi_vessel_name: string
    lopfi_arrival_time: string
    lopfi_departure_time: string
    dipla_location_id: string
    dipla_location_name: string
    dipla_voyage: string
    dipla_service: string
    dipla_vessel_code: string
    dipla_vessel_name: string
    dipla_arrival_time: string
    dipla_departure_time: string
    first_mother_vessel_origin_location_id: string
    first_mother_vessel_origin_location_name: string
    first_mother_vessel_destination_location_id: string
    first_mother_vessel_destination_location_name: string
    first_mother_vessel_voyage: string
    first_mother_vessel_service: string
    first_mother_vessel_code: string
    first_mother_vessel_name: string
    first_mother_vessel_arrival_time: string
    first_mother_vessel_departure_time: string
    last_mother_vessel_origin_location_id: string
    last_mother_vessel_origin_location_name: string
    last_mother_vessel_destination_location_id: string
    last_mother_vessel_destination_location_name: string
    last_mother_vessel_voyage: string
    last_mother_vessel_service: string
    last_mother_vessel_code: string
    last_mother_vessel_name: string
    last_mother_vessel_arrival_time: string
    last_mother_vessel_departure_time: string
    origin_country: string
    origin_cluster: string
    destination_country: string
    destination_cluster: string
    route_direction: string
    booked_ffes: number
    shipment_ffes: number
    is_hazardous: boolean
    is_reefer: boolean
    contractual_customer_scv_code: string
    contractual_customer_name: string
    contractual_customer_concern_code: string
    contractual_customer_concern_name: string
    contractual_customer_country: string
    contractual_customer_cluster: string
    contractual_customer_value_proposition: string[]
    booked_by_scv_code: string
    booked_by_name: string
    booked_by_concern_code: string
    booked_by_concern_name: string
    booked_by_country: string
    consignee_scv_code: string
    consignee_name: string
    consignee_concern_code: string
    consignee_cooncern_name: string
    consignee_country: string
    price_owner_scv_code: string
    price_owner_name: string
    price_owner_concern_code: string
    price_owner_concern_name: string
    price_owner_country: string
    shipper_scv_code: string
    shipper_name: string
    shipper_concern_code: string
    shipper_concern_name: string
    shipper_country: string
    business_unit_code: string
    business_unit_country_name: string
    delivery_agreement_type: any
    commitment: any
  }
  