// src/app/models/container.ts
export interface Container {
    container_id: string
    shipment_id: any
    last_shipment_id: string
    vessel_id: any
    vessel_next_id: any
    location_id: string
    equipment_leasing_contract_id: string
    rkem_booking_number: any
    lease_start_date: string
    lease_end_date: string
    lease_replacement_value_amount: number
    lease_annual_dep_percent: number
    lease_max_deprecation_percentage: number
    in_out_fleet: any
    profile: any
    production_year: string
    in_location: string
    payload_weight: string
    tare_weight: string
    equipment_size: string
    equipment_group: string
    equipment_dimensions: string
    equipment_own_type: string
    equipment_inspected_status: boolean
    equipment_cleaned_status: boolean
    equipment_sound_status: boolean
    equipment_surveyed_status: boolean
    equipment_repair_auth: boolean
    equipment_repair_status: boolean
    equipment_blocked_status: boolean
    equipment_redelivery_status: boolean
    equipment_sell_scrap_status: boolean
    equipment_type: string
    equipment_reefer_type: any
    equipment_type_rock: string
    equipment_manufacturer: any
    equipment_leasing_company: string
    equipment_type_rkem: string
    sub_code: string
    is_reefer: boolean
    is_live_reference: boolean
    transportation_status: string
    current_location_id: string
    current_loc_id: string
    current_pool_id: string
    current_vessel_id: any
    is_empty: boolean
    cargo_weight_kg: number
    cargo_weight_lb: number
    stuffed_hs_commodity_code: string
    stuffed_commodity_name: string
    parent_commodity_name: string
    previous_parent_commodity_name: string
    pre_trip_inspection_status: any
    latest_pti_date: any
    is_latest: any
    total_cost_usd_45_day: any
    total_cost_usd_90_day: any
    total_cost_usd_365_day: number
    total_cost_usd_lifetime: number
    total_cost_with_maersk_parts_usd: any
    container_status: string
    is_under_repair: boolean
    damage_code: any
    current_shop: any
    current_work_order: any
    current_work_order_status: any
    current_repair_estimate_cost: any
    status_description: any
    food_upgrade: boolean
    potential_food_upgrade: boolean
    can_transport_scrap: boolean
    eqpno: string
    move: string
    activity: string
    activity_location_id: string
    activity_timestamp: string
    activity_for: any
    activity_week: string
    move_timestamp: string
    booking_number: any
    load_port_location_id: any
    discharge_port_location_id: any
    discharge_pool_id: any
    discharge_port_estimated_arrival_time: any
    port_of_receipt_location_id: any
    place_of_delivery_location_id: any
    final_discharge_port_location_id: any
    service_line_code: any
    voyage_code: any
    train_code: any
    train_trip_number: any
    railcar_code: any
    slot_vessel: any
    bay_vessel: any
    carrier_code: string
    operator_code: string
    next_voyage_code: any
    location: string
    location_type: string
    site_type: string
    lat: number
    lon: number
    city_id: string
    city: string
    country_id: string
    country: string
    cluster_id: string
    cluster: string
    region_id: string
    region: string
    loc_id: string
    loc: string
    pool_id: string
    pool: string
    days_on_site: number
    bucketed_days_on_site: string
    last_update_timestamp: string
    delivery_service: any
    previous_move: string
    equipment_infleet_year: number
    is_transhipment: string
  }
  